TLI-95 refactored profile
