+++ src/router.ts
@@ -26,7 +26,7 @@ const NotFound = () => import('./views/not-found.vue')
 const Notifications = () => import('./views/notifications.vue')
 const PasswordReset = () => import('./views/pwd-reset.vue')
 const Privacy = () => import('./views/privacy.vue')
-const Profile = () => import('./views/profile.vue')
+const Profile = () => import('./components/user-views/profile.vue')
 const Register = () => import('./views/register.vue')
 const Root = () => import('./views/root.vue')
 const Verify = () => import('./views/verify.vue')
